Add a 10-bands equalizer.
Add a equalizer-10bands element in gstreamer pipeline. Values used in equalizer
are obtained from gconf (/system/mafw/mafw-gst-eq-renderer/band<n> property for
band <n>).

Changing the values in gconf is the way to change the values in equalizer, as
it will react as soon as those changes are done.