Add a initialization script
Current initialization script that comes with MAFW
(/etc/X11/Xsession.post/32mafw) launches all plugins with priority 0, except
mafw-gst-renderer, which is launched with priority -7.

In our case, we want to launch also mafw-gst-eq-renderer with the same
priority, -7.

This script runs before the original one, and launches mafw-gst-eq-renderer
with priority -7. Hence, when the original script is executed,
mafw-gst-eq-renderer is already running, so it is skipped.